# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List all questions
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/parse-pdf:
    post:
      operationId: parsePdfQuestions
      tags: [questions]
      summary: Parse multiple questions from an uploaded PDF using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParsePdfBody"
      responses:
        "200":
          description: Parsed questions from PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsePdfResult"
  /questions/parse-image:
    post:
      operationId: parseQuestionImage
      tags: [questions]
      summary: Parse a question from an uploaded image using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseImageBody"
      responses:
        "200":
          description: Parsed question data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a single question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question's text and choices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /questions/{id}/check:
    post:
      operationId: checkAnswer
      tags: [questions]
      summary: Check if the selected answer is correct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckAnswerBody"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckAnswerResult"
  /questions/{id}/explain:
    post:
      operationId: explainAnswers
      tags: [questions]
      summary: Get AI explanation for why each answer is right or wrong
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI explanation for each choice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainResult"
  /questions/{id}/deep-explain:
    post:
      operationId: deepExplainQuestion
      tags: [questions]
      summary: Get in-depth explanation of the primary principles at play
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deep explanation of underlying principles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeepExplainResult"
  /questions/{id}/chat:
    post:
      operationId: chatAboutQuestion
      tags: [questions]
      summary: Follow-up conversation about a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: AI response to the follow-up message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /questions/stats:
    get:
      operationId: getQuestionStats
      tags: [questions]
      summary: Get study statistics
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Study statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionStats"
...
components:
  schemas:
...
    Choice:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        label:
          type: string
        text:
          type: string
        isCorrect:
          type: boolean
      required:
        - id
        - questionId
        - label
        - text
        - isCorrect
    Question:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
          nullable: true
        questionText:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/Choice"
        answered:
          type: boolean
        answeredCorrectly:
          type: boolean
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - questionText
        - choices
        - answered
        - createdAt
    CreateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        projectId:
          type: integer
        choices:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              text:
                type: string
              isCorrect:
                type: boolean
            required:
              - label
              - text
              - isCorrect
      required:
        - questionText
        - choices
    ParsePdfBody:
      type: object
      properties:
        pdfBase64:
          type: string
          description: Base64 encoded PDF data
        projectId:
          type: integer
      required:
        - pdfBase64
    ParsePdfResult:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        totalParsed:
          type: integer
      required:
        - questions
        - totalParsed
    ParseImageBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64 encoded image data
        projectId:
          type: integer
      required:
        - imageBase64
    CheckAnswerBody:
      type: object
      properties:
        choiceId:
          type: integer
      required:
        - choiceId
    CheckAnswerResult:
      type: object
      properties:
        correct:
          type: boolean
        correctChoiceId:
          type: integer
        selectedChoiceId:
          type: integer
      required:
        - correct
        - correctChoiceId
        - selectedChoiceId
    ExplainResult:
      type: object
      properties:
        explanations:
          type: array
          items:
            type: object
            properties:
              choiceId:
                type: integer
              label:
                type: string
              isCorrect:
                type: boolean
              explanation:
                type: string
            required:
              - choiceId
              - label
              - isCorrect
              - explanation
      required:
        - explanations
    DeepExplainResult:
      type: object
      properties:
        principles:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              howItApplies:
                type: string
            required:
              - name
              - description
              - howItApplies
        summary:
          type: string
      required:
        - principles
        - summary
    ChatMessageBody:
      type: object
      properties:
        message:
          type: string
        conversationHistory:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required:
              - role
              - content
      required:
        - message
    ChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
...
    QuestionStats:
      type: object
      properties:
        totalQuestions:
          type: integer
        answeredQuestions:
          type: integer
        correctAnswers:
          type: integer
        incorrectAnswers:
          type: integer
        accuracyPercent:
          type: number
      required:
        - totalQuestions
        - answeredQuestions
        - correctAnswers
        - incorrectAnswers
        - accuracyPercent
    UpdateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              label:
                type: string
              text:
                type: string
              isCorrect:
                type: boolean
            required:
              - label
              - text
              - isCorrect
      required:
        - questionText
        - choices
...