# lib/api-zod/src/generated/types/updateQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateQuestionBodyChoicesItem } from "./updateQuestionBodyChoicesItem";

export interface UpdateQuestionBody {
  questionText: string;
  choices: UpdateQuestionBodyChoicesItem[];
}